refactor(shared-log): extract join-warmup module
Pure code motion: move the join-warmup subsystem out of index.ts into
src/join-warmup.ts as JoinWarmupCoordinator<E> (deps-object style per
checked-prune.ts / sync/index.ts precedent).

Moved verbatim (generic R -> E, entry values are never inspected):
- types JoinWarmupSendState, JoinWarmupRetryTimer,
  JoinWarmupScheduledRetryBatch/Cohort/Slot, JoinWarmupScheduledRetries
- const JOIN_WARMUP_SEND_SPACING_MS
- fields _joinWarmupGenerationByTarget, _joinWarmupSendStateByTarget,
  _joinWarmupRetryTimersByTarget, _joinWarmupScheduledRetriesByTarget,
  _repairSweepJoinWarmupGenerationByTarget (public on the coordinator,
  same names; inbound sites become this.joinWarmup.<name>)
- methods getJoinWarmupGeneration, track/untrack/cancelJoinWarmupTimers,
  cancelJoinWarmupTarget, cancelAllJoinWarmupTargets,
  sleepJoinWarmupTracked, scheduleJoinWarmupRetries,
  armJoinWarmupRetrySlot, queueJoinWarmupSend, drainJoinWarmupSends

Re-assignable SharedLog state (_repairLifecycleController,
_repairRetryTimers, _repairMetrics) crosses the boundary only through
late-bound arrow callbacks; lifecycle aliasing is preserved by one
coordinator instance whose reset() re-assigns its own maps at the
existing ctor/open() init sites.

Deviations from the extraction manifest:
- The coordinator is created lazily at first open() (via
  createJoinWarmupCoordinator + ??=) because borsh-deserialized program
  instances never run the constructor (same reason open() re-creates
  _checkedPrune). Reopens keep the same instance, so a still-running
  drain observes reset()'s map swaps via property lookup as before.
- In cancelJoinWarmupTarget the sweep-generation map delete now runs
  after clearRepairSweepOptimisticPeer instead of between the pending
  peer cleanup and it (commutative synchronous map ops on disjoint
  state; onTargetCancelled carries the staying sweep cleanup).
- Per manifest (contra the original task hint),
  RepairSweepOptimisticPeerState and JOIN_WARMUP_RETRY_SCHEDULE_MS stay
  in index.ts; both are owned by staying code.

Ratchet: removed _joinWarmupGenerationByTarget and
_repairSweepJoinWarmupGenerationByTarget from BASELINE (fields left
index.ts). Open question for reviewers: should the ratchet TARGET grow
to include src/join-warmup.ts, whose fields keep generation tokens?
Tests updated to reach the coordinator (log.joinWarmup.<member>); the
queueJoinWarmupSend sinon stub now targets the coordinator instance so
it intercepts both external and internal callers.
